initial folder and files structure

+++ src/controllers/UserController.ts
@@ -1,17 +1,18 @@
-import { getRepository } from "typeorm"
 import { NextFunction, Request, Response } from "express"
 import { User } from "../entities/User"
+import HTTPException from "../exceptions/HttpException"
+import { AppDataSource } from "../utils/data-source"
 
 export class UserController {
 
-    private userRepository = getRepository(User)
+    private userRepository = AppDataSource.getRepository(User)
 
     async all(request: Request, response: Response, next: NextFunction) {
         return this.userRepository.find()
     }
 
     async one(request: Request, response: Response, next: NextFunction) {
-        return this.userRepository.findOne(request.params.id)
+        return this.userRepository.findOne({ where: { id: request.params.id } })
     }
 
     async save(request: Request, response: Response, next: NextFunction) {
@@ -20,7 +21,11 @@ export class UserController {
 
     async remove(request: Request, response: Response, next: NextFunction) {
         let userToRemove = await this.userRepository.findOneBy({ id: request.params.id })
-        await this.userRepository.remove(userToRemove)
+        if (!userToRemove) 
+            throw new HTTPException(400, 'User not found');
+        
+        await this.userRepository.remove(userToRemove);
+        return;
     }
 
 }
